requirements.txt
aiohttp==3.7.4
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
slackclient==2.9.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
yarl==1.6.3
